Remove a generic parameter from Peer
Peer took it's underlying connection as a generic parameter. As a consequence,
the service_map macro also needed this parameter and thus pulls in all network related
stuff, outright breaking it on WASM and bloating the dependency graph elsewhere.

For the price of a Box per peer object, we now hide this behind a trait object,
simplifying the API and making it work on WASM.

By doing it this way, peer is still able to work with all sorts of different
underlying connections, but the type does no longer clutter the type of Peer.