Move Service Trait into macro.
This is necessary since we implement it for the Message types, but because of orphan rules,
either the trait or the type needs to be in the same crate as the impl.

We want to allow users to use Services provided by a library, which means the
type would not be defined in the application crate that calls the macro to set a
namespace. Thus the Service trait needs to be there so we can implement it.

By putting it in the module generated by the macro it is automatically namespaced,
so the Namespace trait has been removed.

I chose not to remove the Service trait, because being able to get the sid from
a type is quite handy and it would have complicated other things, like the
RemoteAddr needs the sid, and the only other place to get it would have been the
Services (service map), but then it would have had to store an Arc to that,
making the RemoteAddr bigger in size, etc.