Clean up the revert.
Conclusion: Don't make service map an actor. Instead of locking the whole hashmap we now only lock the actual receiver in a mutex.

We also disallow changing handlers after starting to use it. This should be fine as we now have supervised actors, so setting a handler
once allows resilience even if the actor panics.